requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
gunicorn>=22.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
beautifulsoup4>=4.12
pydantic>=2.0
pydantic-settings>=2.0
anthropic>=0.40
psycopg2-binary>=2.9
